requirements.txt
# Django Core
Django>=4.2,<5.0
django-environ>=0.11.2
python-decouple>=3.8

# Authentication & Security
django-axes>=6.1.1

# Cache & Session
django-redis>=5.4.0
redis>=5.0.1

# Task Queue
celery>=5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# MPTT for tree structure
django-mptt>=0.14.0

# Payment
requests>=2.31.0

# Image handling
Pillow>=10.1.0

# Forms
django-crispy-forms>=2.1
crispy-tailwind>=0.5.0

# Persian tools
khayyam>=3.0.17
jdatetime>=4.1.1
persian-tools>=0.0.6

# Development
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
